fra_prototype/requirements.txt
Flask
geopandas
pandas
Pillow
pytesseract
transformers
torch
streamlit
requests
sqlalchemy